+++ src/commonMain/kotlin/Time.kt
@@ -1,0 +1,5 @@
+package fi.papinkivi.crap
+
+expect fun currentMillis(): Long
+
+expect fun sleep(millis: Long)