+++ src/commonMain/kotlin/arduino/Shields.kt
@@ -3,9 +3,9 @@ package fi.papinkivi.crap.arduino
 import fi.papinkivi.crap.*
 import fi.papinkivi.crap.module.*
 
-abstract class Shield(protected val controller: Controller, label: String, func: () -> Unit) : Module(label, func)
+abstract class Shield(controller: Controller<*>, label: String, func: () -> Unit) : Module(controller, label, func)
 
-abstract class RelayShield(controller: Controller, label: String, pins: IntProgression, func: () -> Unit)
+abstract class RelayShield(controller: Controller<*>, label: String, pins: IntProgression, func: () -> Unit)
     : Shield(controller, label, func) {
     private val pins = pins.toList()
 
@@ -14,19 +14,15 @@ abstract class RelayShield(controller: Controller, label: String, pins: IntProgr
     private val nextNo get() = relays.size + 1
 
     fun relay(pin: Int = pins[relays.size]) = Relay(
-        "J$nextNo",
+        controller,
+        controller.pinsById.getValue(pin.toString()) as DigitalPin,
         "Relay$nextNo",
-        controller.pinsById.getValue(pin.toString()) as DigitalPin
-    ).apply { relays.add(this) }
-
-    override fun setup() {
-        debug { "Setup relays" }
-        relays.forEach { it.setup() }
-    }
+        "J$nextNo"
+    ).apply { relays.add(controller.attach(this)) }
 }
 
 /** https://docs.arduino.cc/tutorials/4-relays-shield/4-relay-shield-basics */
-class FourRelayShield(controller: Controller)
+class FourRelayShield(controller: Uno)
     : RelayShield(controller, "Relay Shield v2.1", 7 downTo 4, {}) {
     /** Relay 1 */
     val j1 = relay()