+++ src/jvmMain/kotlin/Time.kt
@@ -1,0 +1,6 @@
+package fi.papinkivi.crap
+
+actual fun currentMillis() = System.currentTimeMillis()
+
+actual fun sleep(millis: Long) = Thread.sleep(millis)
+