+++ src/jvmTest/kotlin/arduino/AllInOne.kt
@@ -1,0 +1,8 @@
+package fi.papinkivi.crap.arduino
+
+fun main() = with(Uno(1)) {
+    relayShield()
+    dallasTemperature()
+    switch()
+    printPinOut()
+}