+++ src/jvmTest/kotlin/arduino/Dallas.kt
@@ -1,12 +1,7 @@
 package fi.papinkivi.crap.arduino
 
-fun main() = with (Uno("COM3")) {
-    val device = attachDallasTemperature()
-    var previous = 0f
-    while (connected) {
-        previous = device.celsius.apply {
-            if (previous != this)
-                println("Changed to $this °C.")
-        }
+fun main() = Uno("COM3").setup {
+    dallasTemperature().onChange {
+            old, new -> println("$old -> $new °C")
     }
-}
+}()