+++ src/jvmTest/kotlin/arduino/Relays.kt
@@ -1,13 +1,6 @@
 package fi.papinkivi.crap.arduino
 
-import java.util.concurrent.TimeUnit
-
-fun main() = with (Uno("COM3")) {
-    with(attach4RelayShield()) {
-        while (connected) {
-            relays.shuffle()
-            relays.first().change()
-            TimeUnit.SECONDS.sleep(1)
-        }
-    }
-}
+fun main() = object : Uno(1) {
+    val relays = relayShield().relays
+    override fun loop() { relays.random().change() }
+}()