+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2022-10-12
+### Added
+- Ability to disconnect.
+- Setup and loop phases.
+- Extendable controllers.
+- On-off switch.
+- Free pin search.
+- Added check to avoid use same pin twice.
+- PinOut printing.
+
+### Changed
+- Fixed protocol version number.
+- Fixed boolean reading.
+
 ## [2.0.0] - 2022-10-11
 ### Added
 - Dallas Temperature sensors support.